utilise des interfaces pour séparer les préoccupations
+ un peu de formattage/nettoyage
