+++ src/fr/lomateo/controls/ControllablePersonnage.java
@@ -1,0 +1,15 @@
+package fr.lomateo.controls;
+
+public interface ControllablePersonnage {
+
+	void setSaut(boolean b);
+
+	void setMarche(boolean b);
+
+	void setVersDroite(boolean b);
+
+	void setDxJ(int i);
+
+	void setFrappe(boolean b);
+
+}
