+++ src/fr/lomateo/main/Joueur.java
@@ -1,0 +1,9 @@
+package fr.lomateo.main;
+
+import java.awt.Graphics2D;
+
+public interface Joueur {
+
+	void deplacement(Graphics2D g2);
+
+}
