+++ src/fr/lomateo/personnages/DefaultJoueur.java
@@ -4,95 +4,95 @@ import java.awt.Graphics2D;
 import java.awt.Image;
 
 import javax.swing.ImageIcon;
-
+import fr.lomateo.main.Joueur;
 import fr.lomateo.main.Scene;
 import fr.lomateo.structures.Structures;
 
+public class DefaultJoueur extends DefaultPersonnage implements Joueur {
 
-public class Joueur extends Personnages {
+	private final Scene scene;
+	private final String nom;
 
+	// Variables pour la physique du saut et de la chute
+	private double vitesseSaut = 6.5;
+	private double vitesseStautActuelle = vitesseSaut;
+	private double vitesseDeChuteMax = 4;
+	private double vitesseDeChuteActuelle = 0.1;
 
-    private final Scene scene;
-    private final String nom;
-    
-    //Variables pour la physique du saut et de la chute
-    private double vitesseSaut = 6.5;
-    private double vitesseStautActuelle = vitesseSaut;
-    private double vitesseDeChuteMax = 4;
-    private double vitesseDeChuteActuelle = 0.1;
-    
-    //Variable pour définir le sol d'un joueur
-    private int ySolJoueurs;
-	
-	public Joueur(int x , int y , String nom ,Scene scene){
-		super(x , y , 38 ,100, nom ,scene);
-		
+	// Variable pour définir le sol d'un joueur
+	private int ySolJoueurs;
+
+	public DefaultJoueur(Scene scene, int x, int y, String nom, boolean versDroite) {
+		super(x, y, 38, 100, nom);
+
 		this.scene = scene;
 		this.nom = nom;
 		this.ySolJoueurs = 692;
+		this.versDroite = versDroite;
 	}
-	
-	//Deplacement du joueur
-	public void deplacementJ(Graphics2D g2) {
-		if(!contact(g2)){
+
+	// Deplacement du joueur
+	@Override
+	public void deplacement(Graphics2D g2) {
+		if (!contact(g2)) {
 			this.x += this.dxJ;
-			this.paint(g2);	
-		}else if(contact(g2)){
+			this.paint(g2);
+		} else if (contact(g2)) {
 			this.x += 0;
 			this.marche = false;
 			this.paint(g2);
 		}
-				
+
 	}
-	
-	//Méthode qui vérifie les contacts avec les structures
-	public boolean contact(Graphics2D g2){
-		
-		for(Structures structure : this.scene.structures){
-			
-			//contact horizontal
-			if((this.contactArriere(structure) && !this.isVersDroite()) || (this.contactAvant(structure))){
+
+	// Méthode qui vérifie les contacts avec les structures
+	private boolean contact(Graphics2D g2) {
+
+		for (Structures structure : this.scene.structures) {
+
+			// contact horizontal
+			if ((this.contactArriere(structure) && !this.versDroite) || (this.contactAvant(structure))) {
 				return true;
 			}
-			
-			//contatc vertical
-			if(contactDessous(structure) && this.chute){
-					this.chute = false;
-					this.y = structure.getY() - this.hauteur;
-					this.ySolJoueurs = structure.getY();
-					System.out.println("contact dessou");
-					
-				}else if((!this.isSaut() && this.y + this.hauteur != scene.getYsol()) ){
-					if(this.y + this.hauteur == this.ySolJoueurs){						
-						if(!contactDessous(structure)){
-							System.out.println(contactDessous(structure));
-							this.chute = true;
-						}
+
+			// contatc vertical
+			if (contactDessous(structure) && this.chute) {
+				this.chute = false;
+				this.y = structure.getY() - this.hauteur;
+				this.ySolJoueurs = structure.getY();
+				System.out.println("contact dessou");
+
+			} else if ((!this.saut && this.y + this.hauteur != scene.getYsol())) {
+				if (this.y + this.hauteur == this.ySolJoueurs) {
+					if (!contactDessous(structure)) {
+						System.out.println(contactDessous(structure));
+						this.chute = true;
 					}
 				}
-			if(contactDessus(structure)){
+			}
+			if (contactDessus(structure)) {
 				this.saut = false;
 				this.vitesseSaut = 6.5;
 				this.chute = true;
 			}
-			
+
 		}
 		return false;
 	}
 
-	//methode pour peindre les joueurs
-	public void paint(Graphics2D g2) {
+	// methode pour peindre les joueurs
+	private void paint(Graphics2D g2) {
 		if (this.saut || this.chute) {
 			g2.drawImage(this.saut("personnage" + nom), this.x, this.y, null);
-		} else if (this.isFrappe() && !this.isMarche()) {
-			g2.drawImage(this.coup("personnage" + nom ), this.x, this.y, null);
+		} else if (this.frappe && !this.marche) {
+			g2.drawImage(this.coup("personnage" + nom), this.x, this.y, null);
 		} else {
 			g2.drawImage(this.marche("personnage" + nom, 30), this.x, this.y, null);
 		}
 	}
-	
-	//methode pour la marche
-	public Image marche(String nom, int frequencePas) {
+
+	// methode pour la marche
+	private Image marche(String nom, int frequencePas) {
 		String str;
 		ImageIcon ico;
 		Image img;
@@ -128,9 +128,9 @@ public class Joueur extends Personnages {
 		img = ico.getImage();
 		return img;
 	}
-	
-	//methode pour le saut
-	public Image saut(String nom) {
+
+	// methode pour le saut
+	private Image saut(String nom) {
 		ImageIcon ico;
 		Image img;
 		String str = null;
@@ -163,11 +163,11 @@ public class Joueur extends Personnages {
 
 			if (vitesseDeChuteActuelle < vitesseDeChuteMax) {
 				vitesseDeChuteActuelle += .1;
-			} 
-			
+			}
+
 			else if (this.y + this.hauteur >= this.scene.getYsol()) {
 				this.chute = false;
-				this.y = this.scene.getYsol() -  this.hauteur;
+				this.y = this.scene.getYsol() - this.hauteur;
 			}
 		}
 
@@ -180,7 +180,7 @@ public class Joueur extends Personnages {
 	}
 
 	// methode pour les coups
-	public Image coup(String nom) {
+	private Image coup(String nom) {
 		String str = null;
 		ImageIcon ico;
 		Image img;
@@ -206,5 +206,4 @@ public class Joueur extends Personnages {
 		img = ico.getImage();
 		return img;
 	}
-	
 }
